enable jwt

+++ app/modules/users/controllers.ts
@@ -1,10 +1,11 @@
+import * as jwt from "jsonwebtoken";
 import { Request, Response, NextFunction } from "express";
 import { UserTypes as T } from "./types/users";
+import { tokenizeUser, verifyToken } from "../../utils/tokenHelper";
 import Users from "../../models/users";
 
 const verifyUserPayload = (req: Request, res: Response, next: NextFunction) => {
   const { name, lastName, email, phoneNumber } = req.body;
-
   if (!name || !lastName || !email || !phoneNumber) {
     return res.status(400).json({
       result: "error",
@@ -15,18 +16,45 @@ const verifyUserPayload = (req: Request, res: Response, next: NextFunction) => {
   return next();
 };
 
-const findUsers = async (req: Request, res: Response) => {
-  let query = {};
+const tokenValidation = (req: Request, res: Response, next: NextFunction) => {
+  const token = req.body.token || req.query.token;
 
-  const { name, lastName, email, phoneNumber } = req.query;
-  if (name || lastName || email || phoneNumber) {
-    query = { ...req.query };
+  if (!token) {
+    return res.status(400).json({
+      result: "error",
+      message: "incomplete user information",
+      details: ""
+    });
   }
 
-  let usersFound: Array<T.User>;
+  try {
+    const userData = verifyToken(token);
+
+    req.body.data = userData;
+  } catch (error) {
+    return res.status(500).json({
+      result: "error",
+      message: "internal error",
+      details: error.name
+    });
+  }
+  return next();
+};
+
+const findUser = async (req: Request, res: Response) => {
+  const { userId, name, lastName, email, phoneNumber } = req.body.data;
+  if (!userId) {
+    return res.status(400).json({
+      result: "error",
+      message: "incomplete user information",
+      details: ""
+    });
+  }
+
+  let userFound: T.User | null;
 
   try {
-    usersFound = await Users.find(query);
+    userFound = await Users.findById(userId);
   } catch (error) {
     return res.status(500).json({
       result: "error",
@@ -35,33 +63,24 @@ const findUsers = async (req: Request, res: Response) => {
     });
   }
 
-  if (usersFound.length === 0) {
+  if (!userFound) {
     return res
       .status(400)
       .json({ result: "error", message: "invalid user", details: "" });
   }
 
-  console.log("users found === ", usersFound);
-
-  const sortUsers = usersFound.sort((a: T.User, b: T.User) => {
-    if (a.lastName < b.lastName) {
-      return -1;
-    }
-    if (a.lastName > b.lastName) {
-      return 1;
-    }
-    return 0;
-  });
-
-  return res.status(200).json({ result: "ok", users: usersFound, details: "" });
+  return res.status(200).json({ result: "ok", data: userFound, details: "" });
 };
 
 const createUser = async (req: Request, res: Response) => {
   const newUser: T.User = req.body;
 
   try {
-    const usersSaved = await Users.create(newUser);
-    return res.status(201).json({ result: "ok", userId: usersSaved._id });
+    const userSaved = await Users.create(newUser);
+
+    const token = tokenizeUser(userSaved);
+
+    return res.status(201).json({ result: "ok", data: token });
   } catch (error) {
     return res.status(500).json({
       result: "error",
@@ -72,16 +91,19 @@ const createUser = async (req: Request, res: Response) => {
 };
 
 const deleteUser = async (req: Request, res: Response) => {
-  const { name, lastName, email, phoneNumber } = req.body;
-
-  let deleteResponse: T.DeleteResponse;
-  try {
-    deleteResponse = await Users.deleteOne({
-      name,
-      lastName,
-      email,
-      phoneNumber
+  // TODO: move this data verification to a middleware
+  const { userId, name, lastName, email, phoneNumber } = req.body.data;
+  if (!userId || !name || !lastName || !email || !phoneNumber) {
+    return res.status(400).json({
+      result: "error",
+      message: "incomplete user information",
+      details: ""
     });
+  }
+
+  let deleteResponse: T.DeleteResponse | null;
+  try {
+    deleteResponse = await Users.deleteOne({ _id: userId });
   } catch (error) {
     return res.status(500).json({
       result: "error",
@@ -101,4 +123,4 @@ const deleteUser = async (req: Request, res: Response) => {
     .json({ result: "ok", userId: "", details: "user deleted" });
 };
 
-export { createUser, deleteUser, findUsers, verifyUserPayload };
+export { verifyUserPayload, createUser, tokenValidation, deleteUser, findUser };
